feat: download dialog

+++ call-for-stratagems/app/src/main/java/indie/wistefinch/callforstratagems/utils/DownloadService.kt
@@ -1,5 +1,6 @@
 package indie.wistefinch.callforstratagems.utils
 
+import android.util.Log
 import kotlinx.coroutines.*
 import okhttp3.*
 import java.io.File
@@ -7,59 +8,75 @@ import java.io.FileOutputStream
 import java.io.InputStream
 
 class DownloadService(
-    private val url: String,
-    private val onProgress: (downloaded: Long, total: Long) -> Unit = { _, _ -> },
-    private val onComplete: () -> Unit = {},
-    private val onError: (Exception) -> Unit = { _ -> }
+    var onProgress: (downloaded: Long, total: Long) -> Unit = { _, _ -> },
+    var onComplete: () -> Unit = {},
+    var onError: (Exception) -> Unit = { _ -> }
 ) {
     private val client = OkHttpClient()
     private var job: Job? = null
 
-    fun downloadToFile(scope: CoroutineScope, path: String) {
+    fun downloadToFile(url: String, path: String, scope: CoroutineScope) {
         job = scope.launch(Dispatchers.IO) {
             try {
-                val request = Request.Builder().url(url).build()
-                client.newCall(request).execute().use { response ->
-                    if (!response.isSuccessful) {
-                        throw Exception("Err: Code ${response.code}")
+                val file = File(path)
+                if (!file.getParentFile()?.exists()!!) {
+                    file.getParentFile()?.mkdirs()
+                }
+                if (file.exists()) {
+                    withContext(Dispatchers.Main) {
+                        if (isActive) onComplete()
                     }
+                } else {
+                    val request = Request.Builder().addHeader("Accept-Encoding", "*").url(url).build()
 
-                    val body = response.body ?: throw Exception("Err: Body empty")
-                    val totalLength = body.contentLength()
+                    client.newCall(request).execute().use { response ->
+                        if (!response.isSuccessful) {
+                            throw Exception("Err: Code ${response.code}")
+                        }
 
-                    val file = File(path)
-                    val outputStream = FileOutputStream(file)
-                    val inputStream: InputStream = body.byteStream()
+                        val body = response.body ?: throw Exception("Err: Body empty")
+                        val totalLength = body.contentLength()
 
-                    val buffer = ByteArray(8 * 1024)
-                    var bytesRead = 0
-                    var downloaded = 0L
-                    var lastUpdateTime = System.currentTimeMillis()
+                        val outputStream = FileOutputStream(file)
+                        val inputStream: InputStream = body.byteStream()
 
-                    inputStream.use { input ->
-                        outputStream.use { output ->
-                            while (isActive && input.read(buffer).also { bytesRead = it } != -1) {
-                                output.write(buffer, 0, bytesRead)
-                                downloaded += bytesRead
+                        val buffer = ByteArray(8 * 1024)
+                        var bytesRead = 0
+                        var downloaded = 0L
+                        var lastUpdateTime = System.currentTimeMillis()
 
-                                val currentTime = System.currentTimeMillis()
-                                if (currentTime - lastUpdateTime >= 500 || downloaded == totalLength) {
-                                    onProgress(downloaded, totalLength)
-                                    lastUpdateTime = currentTime
+                        inputStream.use { input ->
+                            outputStream.use { output ->
+                                while (isActive && input.read(buffer).also { bytesRead = it } != -1) {
+                                    output.write(buffer, 0, bytesRead)
+                                    downloaded += bytesRead
+
+                                    val currentTime = System.currentTimeMillis()
+                                    if (currentTime - lastUpdateTime >= 500 || downloaded == totalLength) {
+                                        withContext(Dispatchers.Main) {
+                                            onProgress(downloaded, totalLength)
+                                        }
+                                        lastUpdateTime = currentTime
+                                    }
                                 }
                             }
                         }
+                        withContext(Dispatchers.Main) {
+                            if (isActive) onProgress(downloaded, totalLength)
+                            if (isActive) onComplete()
+                        }
                     }
-                    if (isActive) onProgress(downloaded, totalLength)
-                    if (isActive) onComplete()
                 }
+
             } catch (e: Exception) {
-                if (isActive) onError(e)
+                withContext(Dispatchers.Main) {
+                    if (isActive) onError(e)
+                }
             }
         }
     }
 
-    suspend fun downloadAsStr(): String {
+    suspend fun downloadAsStr(url: String): String {
         var body = ""
         withContext(Dispatchers.IO) {
             val client = OkHttpClient()
